requirements.txt
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
eval_type_backport==0.2.0
filelock==3.16.1
filetype==1.2.0
fonttools==4.55.0
fsspec==2024.10.0
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.49
scipy==1.14.1
setuptools==70.2.0
simsimd==6.0.5
six==1.16.0
stringzilla==3.10.10
sympy==1.13.1
torch==2.6.0.dev20241101
torchaudio==2.5.0.dev20241101
torchvision==0.20.0.dev20241101
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
